requirements.txt
fastapi==0.112.1
passlib==1.7.4
pydantic==2.8.2
PyJWT==2.8.0
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.20
